make scoreview not limited to qualifiers
